Feat: TicketList 컴포넌트 생성 및 수강권 페이지 폴더/파일명 변경
