+++ src/routes/tickets/centerTicket/createTicket.tsx
@@ -1,12 +1,12 @@
 import React, { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import SubHeader from '../../components/common/SubHeader';
-import { LessonTypeEnum, TermUnitEnum } from '../../enums/Ticket';
-import Select from '../../components/common/Select/Select';
-import Input from '../../components/common/Input/Input';
-import { axiosInstance } from '../../utils/apiInstance';
-import { ReactComponent as Plus } from '../../assets/icons/Plus.svg';
-import { ReactComponent as Minus } from '../../assets/icons/Minus.svg';
+import SubHeader from '../../../components/common/SubHeader';
+import { LessonTypeEnum, TermUnitEnum } from '../../../enums/Ticket';
+import Select from '../../../components/common/Select/Select';
+import Input from '../../../components/common/Input/Input';
+import { axiosInstance } from '../../../utils/apiInstance';
+import { ReactComponent as Plus } from '../../../assets/icons/Plus.svg';
+import { ReactComponent as Minus } from '../../../assets/icons/Minus.svg';
 
 const CreateTicket = () => {
   interface StateType {
