+++ src/routes/tickets/issuedTicket/allTicketList.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import SubHeader from '../../../components/common/SubHeader';
+import TicketList from '../../../components/tickets/TicketList';
+
+const AllTicketList = () => (
+  <>
+    <SubHeader title="수강권 부여" />
+    <TicketList />
+  </>
+);
+export default AllTicketList;
